feat: extract typescript types
