+++ src/service/index.ts
@@ -1,31 +1,16 @@
 import axios from 'axios'
-
-type TGetParams = {
-  limit: number | undefined,
-  offset: number | undefined
-}
-
-type TGetItems = (url: string, params?: TGetParams) => Promise<any>
-
-type TGetItemByID = (url: string, id: string) => Promise<any>
-
-type TGenre = {
-  _id: string,
-  name: string,
-  description: string,
-  country: string,
-  year: number,
-}
+import { TGenreObj } from '../modules/genres/genres'
+import { TGetItems, TGetItemByID } from './service'
 
 export const getItems: TGetItems = async (url, params) => {
   const { data: { items } } = await axios.get(url, { params })
 
-  return items.map((data: TGenre) => ({ ...data, id: data._id }))
+  return items.map((item: TGenreObj) => ({ ...item, id: item._id }))
 }
 
 export const getItemById: TGetItemByID = async (url, id) => {
   const composedUrl = url + '/' + id
-  const { data } = await axios.get(composedUrl)
+  const { data: item } = await axios.get(composedUrl)
 
-  return { ...data, id: data._id }
+  return { ...item, id: item._id }
 }
