+++ src/service/service.d.ts
@@ -1,0 +1,7 @@
+type TGetParams = {
+  limit: number | undefined,
+  offset: number | undefined
+}
+
+export type TGetItems = (url: string, params?: TGetParams) => Promise<any>
+export type TGetItemByID = (url: string, id: string) => Promise<any>
